Move ListDropdown into separate module
This is used in many places now, so should move out of the binding module.

https://jira.coreos.com/browse/CONSOLE-938
